Add service that can squash and rebase into merge request refs
The existing `MergeRequests::MergeToRefService` can only create merge
commits. For merge trains and pipelines for merged results, we need a
service that can prepare the ref to be exactly as the target branch
would be if merged. This means that they need both squash and rebase
support. To do so, we introduce a new service
`MergeRequests::CreateRefService`, that combines a new RPC,
`UserRebaseToRef`, with existing RPCs to create a merge request ref
according to the project's merge preferences.