Remove SBoM ingestion transaction
The `sbom_components`, `sbom_component_versions`, and `sbom_sources`
tables are very high contention. This means that it's not necessary to
perform rollbacks on these records if we fail at some point during
insertion, as they'll most likely be needed for something else anyway.
Removing the transaction prevents query timeouts when we have several
transactions blocking on each other.