Show message acknowledging deployment approval
It can be unclear to users whether or not their approval or rejection of
a deployment has taken effect due to a mix of polling delays and quick
loading times.

To get around this, we toast a small message acknowledging that their
approval or rejection of the deployment had executed _successfully_.

We do not show a toast in the event of an error, but the error message
is put into an alert at the top of the page.

This covers both the environment details page and environments page.

Changelog: added
EE: true