Implement frequently visited items container
This component implements the behaviour common to frequently visited
groups and projects, that will be shown in the search modal's default
view.

It reuses the same backing store as the existing context switcher in the
navigation sidebar, so the two should show the same information.

One difference is that this implementation re-reads what's in local
storage each time it's shown, whereas the context switcher does not.

The next commit will implement the project- and group-specific
components, which will use this component.