Fix merge commit result in MergeRequests::CreateRefService
A bug slipped in during the last round of reviews when introducing the
service. The ref was created correctly, but the incorrect result was
returned due to assigning the result of the merge to `source_sha`
instead of `commit_sha`. This caused the merge train pipeline to use the
wrong commit SHA.

This bug is hidden behind a feature flag, and was detected during
testing on staging. No users are affected.

A test gap is what allowed the bug to slip past. We were not testing the
return values of `MergeRequests::CreateRefService#execute`. Additional
checks have been added to the spec that would have caught this bug, and
similar ones for other merge methods.