Show inherited group members when flag is number
For some reason, while I was testing locally, the inheritance type would
always be a string, but in production, we see many integer values being
returned.

So, we convert the flag to a string always before we check its value
with the constant to eliminate the type issue.

Changelog: fixed
EE: true