Surface error when need name exceeds length
Currently when a needed job name exceeds the limit (128 chars)
there is no validation error. The pipeline fails to be created
without any error messages.  This MR adds a validation check
to ensure that the needed job name does not exceed the limit
and surfaces that error message.
Note that a job name can be expanded if it's a parallel job
so adding a validation to the Need entry class is not sufficient.

Changelog: fixed