Add streaming to Anthropic client
This adds the option to stream data from Anthropic by calling the
`stream` method. We added a new method instead of adding an option to
`complete` because streaming does not work with the concept of
exponential backoff.

It also adds the `anthropic-version` header with '2023-06-01' as
Anthropic changed the way streaming works and the previous version will
be deprecated. The old version accumulated the response in the stream,
while the new version only sends the next token.

Changelog: changed
EE: true