Instrument the command palette
This adds Snowplow tracking to several parts of the command palette.

The general idea is to have a correspondence with existing entries in
the new navigation sidebar/context switcher, where possible. The only
difference should be the action, which is `click_command_palette_item`
instead of `click_menu_item`.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/417996.