Handle code suggestions add-on purchase in licenses
When a license for an Online Cloud subscription is used as the current
license and it contains a code suggestions add-on purchase, the needed
data for the code suggestions add-on purchase will now be created. This
happens when an activation code is used for such a license or when a
license seat link is triggered for one.

Additionally, when a license is destroyed, the code suggestions add-on
purchases are either expired, created or updated accordingly to the new
current license.

Changelog: added
EE: true