gitaly: Update server version to include memory leak fix
We have recently started to hit a memory leak issue in production
systems. This bug causes us to build up memory up to the point where
Gitaly becomes completely unresponsive and unable to serve traffic. The
issue was root caused to a change in Gitaly where we have introduced a
new adaptive limiter. This change was thus reverted.

Upgrade the Gitaly server version to 5a12604c5 (Merge branch
'pks-revert-adaptive-limiter' into 'master', 2023-08-17), which is the
latest commit containing the revert.

Changelog: fixed