Do not include min_access_level for admins
In the GitLab for Jira app, admins should be able to view and link
all groups on the instance, even when they are not direct
maintainers of that group.

This change moves the logic to set `min_access_level` from the
backend helper to frontend to allow omitting this param when
user is an admin.

Changelog: changed