Fix AutoDevOps for projects that configured license scanning job
The license scanning job will no longer produce any reports and
will be allowed to fail. To be transparent, it outputs a reference to
the removal of the license scanning artifact based approach. In
cases where a project had configured the license scanning job
without a `script` step, the complete template removal resulted
in a broken pipeline state. Re-adding the job fixes this while making
it visible that the job is no longer producing reports.