Respect merge method and squash preferences for merge train ref
Completes the rollout of
https://gitlab.com/gitlab-org/gitlab/-/issues/420161.

With this change, merge train ref creation respects the merge request's
squash preferences, and the project's merge method preference. For
fast-forward and semi-linear merges, an automatic rebase is performed
for cars 2+ in the train. This is part of enabling fast-forward and
semi-linear merges for merge trains (but does not enable it on its own).

Changelog: changed