Mention @cluster includes id number in path
There has been some customer frustration around how to convert the
`@cluster` path to the `@hashed` path expected by rails.

Add a note to the troubleshooting docs explaining that Praefect's
internal id can be found using the last element of the path, which can
be provided to the `praefect metadata` command to retrieve the `@hashed`
path.