Support explain code for blobs
If the current user is viewing some code blob,
the user should be able to ask the chat to
explain the code.

We will inject the blob's code into the zeroshot
executor's prompt and ask the LLMs to directly
explain the code when instructed.

To make that possible, we will make use of
the Referer header to detect if a user is viewing
a Blob. The referer url will be added as an option
to be extracted by CompletionWorker.

CompletionWorker will then attempt to resolve
and authorize the blob pointed to by the referer url.
If the blob is found and authorized,
it will be available as a context attribute, 'extra_resource'.

The zeroshot executor can then use the attribute to
include the code blob and additional prompt.

The change is guarded with a feature flag.