Fixes failed render when job stage data is missing
This change fixes a Vue render error that prevents diplaying the jobs
table when its stage is missing.

Some jobs may not have a stage due to them being generated as seed data,
as the stage is nullable in our API, which should account for it to be
"null".

Changelog: fixed