Check Arkose's status, bypass captcha on sign-in if there is an outage
- If there is an error with Arkose's API, log the error, check Arkose's
status to see if there is an outage. If it's operational, then show
the error message to the user, otherwise let the user bypass arkose
and sign-in.
- Remove variable showArkoseFailure from SignInArkoseApp in Vue
since it's not used anymore
- Add a variable to the Vue app, arkoseChallengeByPassed
when the challenge is bypassed due to arkose initialization error
- Create a Arkose::StatusService and which calls Arkose's status
API and consider arkose as operational if there is no/minor outage.
- Call Arkose::StatuService after the API call to
::Users::CaptchaChallengeService to avoid sending a status
request even when we don't need to

EE: true
Changelog: added