Add optional client_subscription_id
This adds an optional client_subscription_id to the ai_completion_response
subscription. In addition to that it fixes the GraphqlTriggers to be
able to deal with optional subscription arguments.

This prepares us to allow listening only to a specific client_subscription_id
on the websocket, and to only broadcast messages based on a user_id. This
is important for the chat as well as other aiActions.

This has no breaking changes, nor changes how the subscription gets used.

Changelog: changed
EE: true