Update group path updates with NPM packages
NPM packages not following the naming convention will no longer block
any group path update.

In addition, when the group being updated is a subgroup, NPM packages
will no longer be considered for the update validation. In this case, a
subgroup update has no impact on the GitLab NPM package registry.

In other words, a group path update can be rejected by NPM packages if
and only if:
* The group is a root group.
* That group has at least one NPM package that follows the naming
convention (the scope of the package is the same as the group path).

Changelog: fixed