Track merge train ref merge params
We are in the progress of implementing merge train merges as
fast-forward merges directly from the merge train ref. At the time of
merge, we need to update the `merge_commit_sha` and `squash_commit_sha`
attributes of the merge request. Since the preferences used to create
the ref may have changed in the meantime, we need to persist the
necessary values when creating the ref.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/422483