Fix timeout on group and instance security dashboards
The scanners data is only used on the project security dashboard, but
the query is still executed on the group and instance security
dashboards. The query performs very poorly for groups because it has to
traverse the group hierarchy and find all unique scanners (potentially
hundreds of thousands). In some cases, this can prevent the
vulnerability report from loading. Remove this query since it is
expensive and not needed.

Changelog: fixed
EE: true