fix(422965): Improve docs for flux tutorial
The manually created Secret with the name `gitlab-agent-token` is
blocking the Helm release, because the Helm chart tries to create a
new Secret with the same name.

Helm complains the resource is already existing, but is not managed
by Helm (labels are missing).

As a workaround, I rename the initial Secret. The Helm release will
read the values of the initial Secret `gitlab-agent-token-initial` and
will compose a new Helm-managed Secret with these values.