Update the dependency proxy for Maven packages
So that when the package file is found, its digest (etag) is verified
and returned if all is ok.

If the current user can't interact with the package registry (not
enough permissions), the file from the remote registry is returned.

Example: if the requested file is not in the package registry and the
user can't create packages, then the file from the remote registry is
simply returned.

Also, permissions updated so that:
* anonymous users are rejected (at least guest is required).
* for private projects, we require at least reporter.

This is to follow/be coherent with the existing permissions set for the
dependency proxy for container images and the package registry