Stub super_sidebar_logged_out feature flag locally
We currently stub the `super_sidebar_logged_out` globally and in all
specs, the old navigation will be used for anonymous use. With this
commit we move the stubbing to specs that depend on the old behavior.

This will give us a clear list of specs which we need to fix.

For more context, see:
 - https://gitlab.com/groups/gitlab-org/-/epics/11110
 - https://gitlab.com/gitlab-org/gitlab/-/issues/420119