Introduce no_super_sidebar trait in specs
We are planning to make the super sidebar navigation the default
navigation. Unfortunately we currently rely on the _old_ navigation
behavior in more than 150 spec files. Currently the _old_ navigation is
used in all specs. The plan is to:

1. Introduce a user specific trait (this commit) and move from a global
   "old"-nav switch to a per-test switch.
2. Remove the global feature flag toggle
   (https://gitlab.com/gitlab-org/gitlab/-/merge_requests/126175)
3. Fix all tests to work with the new nav:
   https://gitlab.com/gitlab-org/gitlab/-/issues/420121
4. Remove the trait.

This commit additionally adds the trait to a few specs in order to show
that it doesn't break anything:
 - a few navigation related specs (`navbar_spec.rb`)
 - a few fixtures which otherwise break jest specs
 - a few shared examples and related specs