Remove support for deprecated: { reason: :alpha }
This removes support in our GraphQL deprecations for creating an alpha/
experimental schema item through:

```
deprecated: { reason: :alpha, ... }
```

Instead we now just support this method:

```
alpha: { ... }
```

The former was the original way that alpha schema was introduced.

For a number of months, we've had support for the `alpha:` keyword and
documentation has always reflected this method.

It converts some original uses of the `deprecated: { reason: :alpha }`
style to `alpha:`.