feat: Add new filter option `runner_type` to jobs page in admin area
Currently, the only available filter option considers the job status.
This is not enough for certain organizations. This MR adds the filter
option `runner_type` to the jobs admin page.
This filter is very useful as you want to see the jobs execucted by
a certain runner type.

Changelog: changed