Use Rails 7.1 default for AS::TimeWithZone.name
This removes the following deprecation warning:

DEPRECATION WARNING: ActiveSupport::TimeWithZone.name
has been deprecated and from Rails 7.1 will use the
default Ruby implementation.

This warning is being emitted by calls in graphql-ruby
which has already been updated to support both the
old behaviour and the Rails 7.1 default behaviour in
https://github.com/rmosolgo/graphql-ruby/pull/3774
so we should be safe to update to the new default.