Cleanup FE doc landing page
A lot of the titles in the landing pages are just pages
that we link in the top level navigation anyway. We keep
only the ones that feel as a good "entry" level source.
We are also replacing two pages "development_process" and
"principles" with just the landing page and "getting_started".