spec: Fix tests to not consider dots in identities as crud anymore
When reading both `user.name` and `user.email`, Git considers a certain
set of characters as crud that will be stripped both from the head and
tail of the respective values. This behavior has changed in 1c04cb0744
(ident: don't consider '.' a crud, 2023-08-02) such that dots are not
considered crud anymore. The motivation is e.g. to allow for names such
as "John Doe, Jr." to be accepted without modification.

Some of our tests use dot to denote invalid username. Change these tests
to use ',' instead. This currently blocks the update to Git v2.42 on
Gitaly.