Drop bridge jobs on unknown failures
Sometimes the underlying systems fail to respond and
an error from them would leave the bridge in the running
state undefinetly. Since we don't know the full list of
errors that could cause this, we can catch everything
and let the user decide if they want to retry or not.

Changelog: fixed