Remove pages_deployments[project_id, path_prefix] unique index
Currently we keep pages deployments for about 30m after creating a new
one. This is to avoid racing conditions between the time the new
deployment is created and GitLab Pages is updates its caches with the
latest deployment.

Therefore, while we use this mechanism, we cannot enforce uniqueness on
pages_deployments for project_id and path_prefix, as more than one
deploy for given path_prefix might exist.

Related to:

- Issue that introduced the index:
  - https://gitlab.com/gitlab-org/gitlab/-/issues/416488
- Problem was found while working on
  - https://gitlab.com/gitlab-org/gitlab/-/merge_requests/129534
  - https://gitlab.com/gitlab-org/gitlab/-/issues/416492
Changelog: added