Use `License.feature_available?` to check Usage Ping features
With moving more feature to User Registration (a.k.a Usage Ping)
features we need a SSoT method that can answer that.

The License.feature_available? now fallsback to checking if
the feature is available as Usage Ping feature, when setting
is enabled at instance level.

In order to do that we also had to block some features
from being migrated to Usage Ping features because the would
otherwise trigger a cyclical dependency with ApplicationSetting
model and raise a SystemStackError.