Fix MergeRequests::CreateRefService result source_sha
This is always supposed to be the SHA of the head of the source branch
at the time of creation. This misunderstanding was caused by an
incorrect extrapolation made when extending
`MergeRequests::MergeToRefService`, which grabs this value as the second
parent of the merge commit.

Extrapolating this to merge train refs with squashes led to the creation
of pipelines with an incorrect `source_sha`.

The use of `MergeRequests::CreateRefService` is behind a feature flag,
so there is no changelog.

See https://gitlab.com/gitlab-org/gitlab/-/issues/420161