gitaly-test-spawn: Ensure that the gitlab-shell secret exists
Gitaly has recently enhanced its configuration checks to detect more
invalid configurations. One of these checks asserts that the secret
required to connect to Rails actually exists on disk -- if it doesn't,
Gitaly will refuse to start.

This has in fact broken the "gitlab:setup" CI job because we do not
ensure that the secret actually exists. Let's fix that.

Changelog: fixed