spec/snippet: Don't consider dot author as cruft anymore
Git considers specific characters part of the author and committer
signature as "cruft" that is to be stripped or disallowed. The rules of
what is considered a cruft character has changed in Git v2.42.0 though
to not include the dot (".") anymore. One test for snippet repositories
still relies on that behaviour though.

Adapt the test to a different character that Git still considers as
cruft.