Update 'docker' and 'docker:dind' image version in examples
In our documentation about how to use Docker-in-Docker approach
for building container images in the GitLab CI/CD job, we're
explicitly pointing version `20.10.16`.

While pointing a specific version is a very good and encouraged
by us strategy, the version 20.10.16 is now a year old. And as many
users - especially the ones who just start to experiment with GitLab
CI/CD - simply copy examples from our docs and leave them as is, it
seems to be a good idea, to update this refferenced version to
whatever is the newest.