Add username to users mapper
This change updates the get_members graphql query to include the
username in the returned result. It updates the user_mapper with
methods to create and save the source_username and destination_username
as a hash in redis. It adds a usernames pipeline for updating all
issue/mr/note username mentions with the correct username where
necessary.

Changelog: fixed