Change fallback for logged-out sidebar
Currently the logged-out sidebar will fall-back to the "Your Work"
panel. Unfortunately this doesn't make a lot of sense, because the
"Your Work" panel relies on a user object existing. It also leads to
exceptions currently.

Furthermore we add missing specs and insure that all the panels render
fine, even if the user object is nil.