Track merged_commit_sha on merge
This ties together a couple of recent changes to ensure fast-forward
merge trains merges use the correct SHAs on merge, and that the
associated merge request gets updated with the correct values:

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/129311 we added
the column `merged_commit_sha`, which represents the SHA of every merged
merge request, independently of the merge method.

In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/129979, we
added `train_ref` to `merge_requests.merge_params`, which contains the
exact SHAs used to construct the ref, which should be written to the
merge request on merge.

This change is behind the existing feature flag
`fast_forward_merge_trains_support`, see
https://gitlab.com/gitlab-org/gitlab/-/issues/282442