URL parsing and sanitizing logic added for Relay State
Default Relay State is sent from Idp end during SAML SSO process.
This value is used for redirecting user to the subpath specified.
With this change we are making sure the user is only redirected to subpaths in application.
And the subpaths specified are not malicious

Changelog: changed
MR:
EE: true