Create iid sequence for ci_pipelines with new projects
The initialization of the sequence is not atomic and
due to race conditions some pipelines could reinitialize
multiple times, resulting in attempts to insert rows
with duplicate values. This fix moves the initialization
from the first pipeline creation to the project creation,
so the sequence will already be initialized when the
first pipelines are created.

Changelog: fixed