Add diagonal tolerance when moving the mouse towards flyout menu
This is implemented with the help of a SVG element, because we have
to listen to mouse events in two triangle-shaped areas, and also
update these areas as the mouse moves.

The reason we need two is that the flyout can shift and flip position
when there wouldn't be enough screen space otherwise.

Changelog: changed