Add `finding_id` to the `vulnerabilities` table
There is an existing `has_many`/`belongs_to` association between the
`vulnerabilities` and `vulnerability_occurrences` tables.

We want to make this a legit 1-to-1 relationship eventually. As a step
towards that we are adding an `id` column that will eventually be used
for a foreign ey constraint.

after adding the column and updating the app code to write to the
column, we need to backfill the new column. [MR 130058][0] will handle
the backfill

resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/418970

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/130058

Changelog: changed
EE: true