Remove schema from bg template
Background migrations by default use latest if the schema is not set,
see
https://gitlab.com/gitlab-org/gitlab/-/blob/v16.3.0-ee/spec/spec_helper.rb?ref_type=tags#L144.

Using a specific schema could lead to false positives, for example
a column referenced on the background migration could be modified which
would break the specs. This would make the background migration fail if
this one is still in progress. The error will be noticeable at merge
request level if the background migration were to use the 'latest'
schema, and would be silently skipped if the background migration was
using a specific migration number.