Put exact matches first when searching for groups to transfer
When transferring a group to another group, previously
`Groups::AcceptingGroupTransfersFinder#execute` would return a list of
accessible groups ordered by the group's path. Suppose you had a
number of groups with the paths:

1. gitlab-org
2. gitlab-org/a/1
3. gitlab-org/1

Groups 2 and 3 would always be prioritized ahead of 1 since the `ORDER
BY path` would put them ahead. When there are thousands of groups,
that makes it almost impossible to locate the exact match,
`gitlab-org`.

This commit introduces the ability for `fuzzy_search` to return items
that match exactly first when the `exact_matches_first` flag is
enabled.
`Groups::AcceptingGroupTransfersFinder#execute` uses this
parameter selectively for now. We might consider turning this on for
all searches.

When the `include_parents` parameter is set to `false` (the default),
the query plan does not seem to be affected because when a `search`
parameter is available, the `fuzzy_search` `ORDER BY` is used instead
of the default `ORDER BY path ASC, id ASC` query.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/423958

Changelog: fixed