Add migration to sync ScanResultPolicyRead
Security policies are stored as YAML files in the
security policy project. This required to read from git whenever
we need data from a security policy. We now also store this data in the
scan_result_policies table. To sync existing polices to the table,
the migration in this MR kicks of sync jobs for all
security policy configurations.

Changelog: other