Replace context switcher with header in sidebar
Now that the global search modal replicates the functionality of the
context switcher (see
https://gitlab.com/gitlab-org/gitlab/-/issues/409522 and
https://gitlab.com/gitlab-org/gitlab/-/issues/415809 for prior work),
the context switcher can be removed.

In its place, a static header specific to each panel type is shown at
the top of the sidebar. The old context header (which was the toggle for
the context switcher when logged in, and a non-interactive element when
logged out) has largely been replaced by changing the "main" or
"overview" link for the current context (if there is one). It now
displays the current context's name (when appropriate), and shows that
entity's avatar.

This does not remove the frontend code for the context switcher. That
will be done in a follow-up.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/415811.

Changlog: changed