Use safe_format for common link pattern
This transforms a somewhat common pattern of `html_safe`/link
interpolation with the `safe_format`/`tag_pair` helpers.

This was implemented using [Comby][1]. The configuration file used can
be found at
https://gitlab.com/gitlab-org/frontend/pajamas-adoption-scanner/-/blob/safe-format-migration/comby/html_safe.toml.

The tool was run with something like this command in the GitLab
repository:

    comby \
	-config ~/dev/pajamas-adoption-scanner/comby/html_safe.toml \
	-custom-matcher ~/dev/pajamas-adoption-scanner/comby/haml.json \
	-matcher .generic -in-place \
	-f .haml

It may be there are other common patterns of `html_safe` usage that can
be transformed in this way in future iterations.

[1]: https://comby.dev