Add ActivityPub Releases profile and outbox
This MR is the first part in the implementation of ActivityPub in
Gitlab, as specified in [my design
documents](https://gitlab.com/oelmekki/gitlab-activitypub-design), in
the context of [epic 260](https://gitlab.com/groups/gitlab-org/-/epics/260)
, and more specifically
[issue 21582](https://gitlab.com/gitlab-org/gitlab/-/issues/21582).

Summary of those sources: we want to implement the ActivityPub protocol
in Gitlab to allow it to federate instances and join the fediverse, so
that:

* the fediverse (like Mastodon, Lemmy, and apparently soon Threads) can
  subscribe to content in Gitlab
* various instances of Gitlab can communicate
* ultimately, cross instances merge requests become possible

This is the very humble beginning. As a first step of the first step of…
we're implementing a way for the fediverse to read content from Gitlab.

We're starting by exposing project releases, as it will be immediately
useful: when this MR and the next one (allowing to subscribe to the feed
and receive push notifications) will be merged, fediverse users will be
able to be notified on their favorite fediverse app when their favorite
software releases a new version (provided it's on Gitlab, obviously :)
).

This MR alone does not allow that yet, but it allows to display the
profile of the resource (called an actor, in ActivityPub terminology).
More importantly, it creates the fundations on which the rest of
ActivityPub features will be built.