Add refresh assignment worker
This worker refreshes the user assignments records for
the given root_namespace and the code_suggestions add_on.

It goes through all the assignments and checks if the
user assigned is still elligible or not.
If not elligible it removes the assignment to free up
the seat.

EE: true
Changelog: added