Support for include:rules:changes
Adds support for 'rules:changes' to be used
in the CI config with 'include'. This allows users
to conditionally include files based on file
change conditions.  It includes functionality for
subkeys 'paths' and 'compare_to'. To support
this new keyword, the 'pipeline' object is now
passed in through External::Context. FF removed.

Changelog: added