Do not checkout connection from pool in DatabaseSampler
As part of debugging
https://gitlab.com/gitlab-org/gitlab/-/issues/423945#note_1544793516 we
were looking into why Rails is constantly timing out when trying to
close connections to the database. It keeps trying to close a whole pool
of connections until no connections are in use in the pool. We had an
arbitrary timeout waiting for the pool to empty but it would always
timeout (even in local testing).

The reason we sometimes need to disconnect to the datbase is that we
have periodic service discovery looking for the current list of replica
databases. Every time that changes in Rails we need to disconnect any
previous list of replicas.

In the end we tracked down this `DatabaseSampler` which is started in a
Thread when the application boots up. This thread at some point asks
for a `connection` which checks it out of the connection pool. But it
never checks it back in so after startup this connection will forever be
wasted by this Thread. It turns out that this `DatabaseSampler` never
really needed a connection from the pool for the specific things it was
doing so I refactored the code to just avoid calling `.connection` and
it seems to fix this offence.