doc/gitaly: Update configuration to not use `[gitlab-shell]` section
Gitaly is in the process of deprecating the `[gitlab-shell]` section in
its configuration. This section implicitly identified the location of
two different mechanisms that have been historically hosted by the
gitlab-shell project:

    - The `.gitlab_shell_secret` file, which contains the secret that is
      required to connect to the GitLab application.

    - The `custom_hooks` directory, which hosts Git hooks that shall be
      executed whenever changes are performed in a repository hosted by
      Gitaly.

These have instead been replaced by standalone items that are not
specific to the gitlab-shell project anymore. Update our documentation
to reflect the new configuration.