Revert wildcard docs update
While I understand the gist of the idea, it eventually
makes the resulting documentation at least confusing, if
not wrong at all.

All the other sections in the document (gladly!) still refer
to the `examples.io` domain as example pages domain. I'm currently
in the situation that I configured my instance exactly as documented
and slowly realize that this won't work. Mostly because wildcard
certificates for `*.example.com` are common, but not for
`*.pages.examples.com`

Will redraft the documentation in the next commit

This reverts commit d556719a60a829f3d00686948cfb1bf60e547a29.