Provide and edit dismissal reason when dismissing finding
When dismissing a finding in the vulnerability finding
modal, the dismissal reason can be provided. The comment
section has been reworked to become an edit section for
both the reason and comment. The dismissal note actions
have been reworked to account for editing the reason too.

Changelog: added
EE: true