Clear messageChunks buffer on mount
AiGenieChatMessage requires a non-reactive buffer
to store the intermediate chunk messages streamed
from the backend. Why? Because the chunk messages
may be out of order and the frontend should
only display them serially in order.

The non-reactive buffer is an property on a Vue component.
When a vue component is rendered with v-for,
the options properties are cloned and the buffer must
be cleared.
It means if one chat message utilized the buffer already,
the next chat message will have the same content in its buffer on mount.