Support for marking Gitsign certs as Unverified
A number of users are wishing to sign their commits with x509 certificates that do not meet the current GitLab requirements, namely the presence of a crlDistributionPoints or non-empty subject. Commits signed as such do not even show in the GitLab UI as unverified, simply having no signing status next to them. This MR relaxes the restrictions so that those certs are accepted, and instead attempts to move those checks to the verified field.

Changelog: changed