Fix tooltip rendering in user bar
This ensures that GitLab UI is configured *before* the super sidebar is
mounted, so that all tooltips have the `gl-tooltip` class applied. See
https://gitlab.com/gitlab-org/gitlab-ui/-/blob/v66.7.0/src/config.js#L13.

Because the super sidebar exists in a separate chunk, it actually
executes *before* the `main.js` script, even though it's listed later in
the document. This is likely due to webpack's runtime managing the
execution order of modules. See
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/114238#note_1312900193.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/424465.

Changelog: fixed