Include grace period for SM add-on expiration
When creating or updating a `GitlabSubscription::AddOnPurchase` for an
SM instance its expiration date is set to the expiration date of the
current license. While the license itself has a grace period, it is not
applied to the add-on purchase. Similar logic for the same add-on on
SaaS includes the expiration date of the add-on plus a grace period.

This change updates the creation and update to use a grace period on top
of the expiration date. This is achieved by using the current license's
`block_changes_at` attribute which is set to value we are looking for
anyway.

Changelog: changed
EE: true