Refactor experimental features toggle for non-AI features
Currently, top-level ultimate namespaces on SaaS will
allow namespace owners to opt-in to experimental features but
only if AI features are enabled on that namespace.

The documentation at
https://docs.gitlab.com/ee/user/group/manage.html#enable-experiment-features
doesn't suggest that this is an AI only feature so this
commit refactors the toggle to work outside of the context
of AI features.

EE: true
Changelog: changed