Extract safe_url to PORO and update usage
The SafeUrl module expects the objects that include it to have a `url`
method containing a valid URL. This assumption fails in the case of web
hooks when we're using the masking functionality, where masked values
are replaced with `{placeholder}`.

This MR replaces the concern with a PORO that accepts the URL as an
argument. IMO this is a little easier to manage than using a concern,
as we no longer have to satisfy the concern's assumptions about the
shape of the class that uses it.